Add submodule `view`
taken from `tools._plot`
