+++ src/felupe/tools/__init__.py
@@ -4,7 +4,6 @@ from ._newton import NewtonResult
 from ._newton import fun_items as fun
 from ._newton import jac_items as jac
 from ._newton import newtonrhapson
-from ._plot import ViewField, ViewMesh, ViewSolid, ViewXdmf
 from ._post import curve, force, moment
 from ._project import extrapolate, project, topoints
 from ._save import save
@@ -26,8 +25,4 @@ __all__ = [
     "save",
     "solve",
     "NewtonResult",
-    "ViewMesh",
-    "ViewField",
-    "ViewXdmf",
-    "ViewSolid",
 ]
