+++ src/felupe/view/_scene.py
@@ -18,8 +18,6 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
 
-from ..math import displacement, eigvalsh, equivalent_von_mises, tovoigt
-
 
 class Scene:
     """Base class for plotting a static scene.
@@ -290,278 +288,3 @@ class Scene:
             plotter.add_axes()
 
         return plotter
-
-
-class ViewXdmf(Scene):  # pragma: no cover
-    """Provide Visualization methods for a XDMF file generated by
-    `:meth:`Job.evaluate(filename="result.xdmf")`. The warped (deformed) mesh is created
-    from the values of the point-data "Displacement".
-
-    Parameters
-    ----------
-    filename : str
-        The filename of the XDMF file (including the extension).
-    time : float, optional
-        The time value at which the data is extracted (default is 0).
-
-    Attributes
-    ----------
-    mesh : pyvista.UnstructuredGrid
-        A generalized Dataset with the mesh as well as point- and cell-data. This is
-        not an instance of :class:`felupe.Mesh`.
-
-    """
-
-    def __init__(
-        self,
-        filename,
-        time=0,
-    ):
-        "XDMF Result file reader and plotter."
-
-        self.filename = filename
-        self.mesh = self._read(time)
-
-    def _read(self, time):
-        "Read the file and obtain the mesh."
-
-        import pyvista as pv
-
-        self.file = pv.XdmfReader(self.filename)
-        self.file.set_active_time_value(time)
-
-        mesh = self.file.read()[0]
-
-        return mesh
-
-    def set_active_time_value(self, time):
-        "Set new active time value and re-read the mesh."
-
-        self.mesh = self._read(time)
-
-
-class ViewMesh(Scene):
-    """Provide Visualization methods for :class:`felupe.Mesh` with optional given
-    dicts of point- and cell-data items.
-
-    Parameters
-    ----------
-    mesh : felupe.Mesh
-        The mesh object.
-    point_data : dict or None, optional
-        Additional point-data dict (default is None).
-    cell_data : dict or None, optional
-        Additional cell-data dict (default is None).
-    cell_type : pyvista.CellType or None, optional
-        Cell-type of PyVista (default is None).
-
-    Attributes
-    ----------
-    mesh : pyvista.UnstructuredGrid
-        A generalized Dataset with the mesh as well as point- and cell-data. This is
-        not an instance of :class:`felupe.Mesh`.
-
-    """
-
-    def __init__(self, mesh, point_data=None, cell_data=None, cell_type=None):
-        self.mesh = mesh.as_pyvista(cell_type=cell_type)
-
-        if point_data is None:
-            point_data = {}
-
-        if cell_data is None:
-            cell_data = {}
-
-        for label, data in point_data.items():
-            self.mesh.point_data[label] = data
-
-        for label, data in cell_data.items():
-            self.mesh.cell_data[label] = data
-
-        self.mesh.set_active_scalars(None)
-        self.mesh.set_active_vectors(None)
-        self.mesh.set_active_tensors(None)
-
-
-class ViewField(ViewMesh):
-    """Provide Visualization methods for :class:`felupe.FieldContainer`. The warped
-    (deformed) mesh is created from the values of the first field (displacements). By
-    default, the "Deformation Gradient" tensor, the "Logarithmic Strain" tensor and the
-    "Principal Values of Logarithmic Strain" are evaluated as field-related items of the
-    cell-data dict. Optional items of given point- and cell-data overwrite these default
-    field-related cell-data items.
-
-    Parameters
-    ----------
-    field : felupe.FieldContainer
-        The field-container.
-    point_data : dict or None, optional
-        Additional point-data dict (default is None).
-    cell_data : dict or None, optional
-        Additional cell-data dict (default is None).
-    cell_type : pyvista.CellType or None, optional
-        Cell-type of PyVista (default is None).
-    project : callable or None, optional
-        Callable to project internal cell-data at quadrature-points to mesh-points
-        (default is None). Valid callables are :class:`~felupe.project` or
-        :class:`~felupe.tools.extrapolate`.
-
-    Attributes
-    ----------
-    mesh : pyvista.UnstructuredGrid
-        A generalized Dataset with the mesh as well as point- and cell-data. This is
-        not an instance of :class:`felupe.Mesh`.
-
-    See Also
-    --------
-    felupe.project: Project given values at quadrature-points to mesh-points.
-
-    """
-
-    def __init__(
-        self, field, point_data=None, cell_data=None, cell_type=None, project=None
-    ):
-        point_data_from_field = {}
-        cell_data_from_field = {}
-
-        if project is None:
-            cell_data_from_field = {
-                "Deformation Gradient": field.evaluate.deformation_gradient()
-                .mean(-2)
-                .T,
-                "Logarithmic Strain": field.evaluate.strain(tensor=True, asvoigt=True)
-                .mean(-2)
-                .T,
-                "Principal Values of Logarithmic Strain": field.evaluate.strain(
-                    tensor=False
-                )
-                .mean(-2)
-                .T,
-            }
-        elif callable(project):
-            point_data_from_field = {
-                "Deformation Gradient": project(
-                    field.evaluate.deformation_gradient(), field.region
-                ),
-                "Logarithmic Strain": project(
-                    field.evaluate.strain(tensor=True, asvoigt=True), field.region
-                ),
-                "Principal Values of Logarithmic Strain": project(
-                    field.evaluate.strain(tensor=False), field.region
-                ),
-            }
-        else:
-            raise TypeError("The project-argument must be callable or None.")
-
-        point_data_from_field["Displacement"] = displacement(field)
-
-        if point_data is None:
-            point_data = {}
-
-        if cell_data is None:
-            cell_data = {}
-
-        super().__init__(
-            mesh=field.region.mesh,
-            point_data={**point_data_from_field, **point_data},
-            cell_data={**cell_data_from_field, **cell_data},
-            cell_type=cell_type,
-        )
-
-
-class ViewSolid(ViewField):
-    """Provide Visualization methods for :class:`felupe.Field` and `felupe.SolidBody`.
-    The warped (deformed) mesh is created from the values of the first field
-    (displacements). By default, the "Deformation Gradient" tensor, the
-    "Logarithmic Strain" tensor and the "Principal Values of Logarithmic Strain" are
-    evaluated as field-related items of the cell-data dict. Optional items of given
-    point- and cell-data overwrite these default field-related cell-data items.
-
-    Parameters
-    ----------
-    field : felupe.FieldContainer
-        The field-container.
-    solid : felupe.SolidBody or felupe.SolidBodyIncompressible or None, optional
-        A solid body to evaluate the (Cauchy) stress (default is None).
-    stress_type : str, optional
-        The type of stress, either "Cauchy" or "Kirchhoff, which is exported (default is
-        "Cauchy").
-    point_data : dict or None, optional
-        Additional point-data dict (default is None).
-    cell_data : dict or None, optional
-        Additional cell-data dict (default is None).
-    cell_type : pyvista.CellType or None, optional
-        Cell-type of PyVista (default is None).
-    project : callable or None, optional
-        Callable to project stress at quadrature-points to mesh-points (default is
-        None). Valid callables are :class:`~felupe.project` or
-        :class:`~felupe.tools.extrapolate`.
-
-    Attributes
-    ----------
-    mesh : pyvista.UnstructuredGrid
-        A generalized Dataset with the mesh as well as point- and cell-data. This is
-        not an instance of :class:`felupe.Mesh`.
-
-    See Also
-    --------
-    felupe.project: Project given values at quadrature-points to mesh-points.
-    """
-
-    def __init__(
-        self,
-        field,
-        solid=None,
-        stress_type="Cauchy",
-        point_data=None,
-        cell_data=None,
-        cell_type=None,
-        project=None,
-        **kwargs,
-    ):
-        if point_data is None:
-            point_data = {}
-
-        if cell_data is None:
-            cell_data = {}
-
-        point_data_from_solid = {}
-        cell_data_from_solid = {}
-
-        if solid is not None:
-            stress_from_field = {
-                "cauchy": solid.evaluate.cauchy_stress,
-                "kirchhoff": solid.evaluate.kirchhoff_stress,
-            }
-            stress = stress_from_field[stress_type.lower()](field)
-            stress_label = f"{stress_type.title()} Stress"
-
-            if project is None:
-                cell_data_from_solid[stress_label] = tovoigt(stress.mean(-2)).T
-                cell_data_from_solid[f"Principal Values of {stress_label}"] = (
-                    eigvalsh(stress).mean(-2).T
-                )
-                cell_data_from_solid[f"Equivalent of {stress_label}"] = (
-                    equivalent_von_mises(stress).mean(-2).T
-                )
-
-            elif callable(project):
-                point_data_from_solid[stress_label] = project(
-                    tovoigt(stress), solid.field.region
-                )
-                point_data_from_solid[f"Principal Values of {stress_label}"] = project(
-                    eigvalsh(stress), solid.field.region
-                )
-                point_data_from_solid[f"Equivalent of {stress_label}"] = project(
-                    equivalent_von_mises(stress), solid.field.region
-                )
-            else:
-                raise TypeError("The project-argument must be callable or None.")
-
-        super().__init__(
-            field=field,
-            point_data={**point_data_from_solid, **point_data},
-            cell_data={**cell_data_from_solid, **cell_data},
-            cell_type=cell_type,
-            project=project,
-        )
